refactor(steps): make step-item a separate directory
