+++ src/step-item/props.ts
@@ -11,11 +11,6 @@ const props: TdStepItemProps = {
     type: String,
     value: '',
   },
-  /** 自定义组件样式 */
-  customStyle: {
-    type: String,
-    value: '',
-  },
   /** 组件类名，用于设置组件外层元素元素类名 */
   externalClasses: {
     type: Array,
@@ -28,6 +23,11 @@ const props: TdStepItemProps = {
   status: {
     type: String,
     value: 'default',
+  },
+  /** 自定义组件样式 */
+  style: {
+    type: String,
+    value: '',
   },
   /** 子步骤条，仅支持 layout  = 'vertical' 时 */
   subStepItems: {
