+++ src/steps/README.en-US.md
@@ -5,14 +5,14 @@
 
 name | type | default | description | required
 -- | -- | -- | -- | --
-current | String / Number | 0 | \- | N
+current | String / Number | - | \- | N
 default-current | String / Number | undefined | uncontrolled property | N
 current-status | String | process | options：default/process/finish/error | N
-custom-style `v0.25.0` | String | - | \- | N
 external-classes | Array | - | `['t-class']` | N
 layout | String | horizontal | options：horizontal/vertical | N
 readonly | Boolean | false | \- | N
 separator | String | line | options：line/dashed/arrow | N
+style | String | - | \- | N
 theme | String | default | options：default/dot | N
 
 ### Steps Events
@@ -26,9 +26,9 @@ change | `({current: string \| number, previous: string \| number})` | \-
 name | type | default | description | required
 -- | -- | -- | -- | --
 content | String / Slot | '' | \- | N
-custom-style | String | - | \- | N
 external-classes | Array | - | `['t-class', 't-class-content', 't-class-title', 't-class-description', 't-class-extra']` | N
 icon | String / Slot | - | \- | N
-status | String | default | options：default/process/finish/error。Typescript：`StepStatus` `type StepStatus = 'default' \| 'process' \| 'finish' \| 'error'`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/steps/type.ts) | N
-sub-step-items | Array | [] | Typescript：`SubStepItem[]` `interface SubStepItem { status: StepStatus, title: string }`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/steps/type.ts) | N
+status | String | default | options：default/process/finish/error。Typescript：`StepStatus` `type StepStatus = 'default' \| 'process' \| 'finish' \| 'error'`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/step-item/type.ts) | N
+style | String | - | \- | N
+sub-step-items | Array | [] | Typescript：`SubStepItem[]` `interface SubStepItem { status: StepStatus, title: string }`。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/step-item/type.ts) | N
 title | String / Slot | '' | \- | N
