+++ src/steps/__test__/index.json
@@ -1,6 +1,6 @@
 {
   "usingComponents": {
     "t-steps": "../steps",
-    "t-step-item": "../step-item"
+    "t-step-item": "../../step-item/step-item"
   }
 }
