+++ src/steps/_example/status/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-steps": "tdesign-miniprogram/steps/steps",
-    "t-step-item": "tdesign-miniprogram/steps/step-item"
+    "t-step-item": "tdesign-miniprogram/step-item/step-item"
   }
 }
